Note: The `update_study_fk` method was accidentally committed in a previous commit but I’ll write the message here as if it wasn’t.
`ingesters.py`: Updated the `IngesterDocumentClinicalTrial` class and added a `update_study_fk` which updates a foreign-key attribute in a `Study` record linked a record that cannot be uniquely identified. Once the FK is updated the old record is deleted. Calls to this method have been added in the `ingest` method.